SMART_PROJECT/requirements.txt
asgiref==3.3.1
certifi==2020.12.5
Django==3.1.6
django-braces==1.14.0
django-registration-redux==2.9
djangorestframework==3.12.2
gunicorn==20.0.4
psycopg2==2.8.6
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
